+++ Tests/Functional/Controller/UndocumentedController.php
@@ -13,6 +13,9 @@ namespace Nelmio\ApiDocBundle\Tests\Functional\Controller;
 
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 
+/**
+ * @Route(host="api.example.com")
+ */
 class UndocumentedController
 {
     /**